requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
python-dotenv==1.0.0
langdetect==1.0.9
